LayoutTests/imported/w3c/web-platform-tests/navigation-api/precommit-handler/precommitHandler-traverse-expected.txt


PASS precommitHandler for a traverse navigation, reject before commit assert_unreached: committed must not fulfill Reached unreachable code
PASS precommitHandler for a traverse navigation, reject after commit assert_equals: hash after commit expected "#1" but got ""
PASS precommitHandler for a traverse navigation, success assert_equals: hash after commit expected "#3" but got "#1"
